+++ packages/mobile/src/cards/__figma__/AnnouncementCard.figma.tsx
@@ -6,7 +6,7 @@ figma.connect(
   AnnouncementCard,
   'https://www.figma.com/design/k5CtyJccNQUGMI5bI4lJ2g/✨-CDS-Components?node-id=61%3A956',
   {
-    imports: ["import { AnnouncementCard } from '@cbhq/cds-mobile/cards/AnnouncementCard';"],
+    imports: ["import { AnnouncementCard } from '@cbhq/cds-mobile/cards/AnnouncementCard'"],
     props: {
       showtopdivider29390: figma.boolean('show top divider'),
       illustration5960: figma.instance('illustration'),