+++ packages/mobile/src/controls/__figma__/Select.figma.tsx
@@ -12,8 +12,8 @@ figma.connect(
   'https://www.figma.com/design/k5CtyJccNQUGMI5bI4lJ2g/✨-CDS-Components?node-id=247-13005&m=dev',
   {
     imports: [
-      "import { Select } from '@cbhq/cds-mobile/controls/Select';",
-      "import { SelectOption } from '@cbhq/cds-mobile/controls';",
+      "import { Select } from '@cbhq/cds-mobile/controls/Select'",
+      "import { SelectOption } from '@cbhq/cds-mobile/controls'",
     ],
     props: {
       startNode: figma.boolean('show start', {