+++ packages/mobile/src/layout/__figma__/Fallback.figma.tsx
@@ -7,7 +7,7 @@ figma.connect(
   Fallback,
   'https://www.figma.com/design/k5CtyJccNQUGMI5bI4lJ2g/✨-CDS-Components?node-id=731-14951&m=dev',
   {
-    imports: ["import { Fallback } from '@cbhq/cds-mobile/layout/Fallback';"],
+    imports: ["import { Fallback } from '@cbhq/cds-mobile/layout/Fallback'"],
     props: {
       shape: figma.enum('shape', {
         circle: 'circle',