+++ packages/mobile/src/navigation/__figma__/TopNavBar.figma.tsx
@@ -18,12 +18,12 @@ figma.connect(
       type: 'title + subtitle',
     },
     imports: [
-      "import { TopNavBar } from '@cbhq/cds-mobile/navigation/TopNavBar';",
-      "import { Divider } from '@cbhq/cds-mobile/layout/Divider';",
-      "import { VStack } from '@cbhq/cds-mobile/layout/VStack';",
-      "import { HStack } from '@cbhq/cds-mobile/layout/HStack';",
-      "import { NavigationTitle } from '@cbhq/cds-mobile/navigation/NavigationTitle';",
-      "import { NavigationSubtitle } from '@cbhq/cds-mobile/navigation/NavigationSubtitle';",
+      "import { TopNavBar } from '@cbhq/cds-mobile/navigation/TopNavBar'",
+      "import { Divider } from '@cbhq/cds-mobile/layout/Divider'",
+      "import { VStack } from '@cbhq/cds-mobile/layout/VStack'",
+      "import { HStack } from '@cbhq/cds-mobile/layout/HStack'",
+      "import { NavigationTitle } from '@cbhq/cds-mobile/navigation/NavigationTitle'",
+      "import { NavigationSubtitle } from '@cbhq/cds-mobile/navigation/NavigationSubtitle'",
     ],
     props: {
       title: figma.string('↳ title'),
@@ -94,11 +94,11 @@ figma.connect(
       type: 'dropdown',
     },
     imports: [
-      "import { TopNavBar } from '@cbhq/cds-mobile/navigation/TopNavBar';",
-      "import { Divider } from '@cbhq/cds-mobile/layout/Divider';",
-      "import { VStack } from '@cbhq/cds-mobile/layout/VStack';",
-      "import { HStack } from '@cbhq/cds-mobile/layout/HStack';",
-      "import { NavigationTitleSelect } from '@cbhq/cds-mobile/navigation/NavigationTitleSelect';",
+      "import { TopNavBar } from '@cbhq/cds-mobile/navigation/TopNavBar'",
+      "import { Divider } from '@cbhq/cds-mobile/layout/Divider'",
+      "import { VStack } from '@cbhq/cds-mobile/layout/VStack'",
+      "import { HStack } from '@cbhq/cds-mobile/layout/HStack'",
+      "import { NavigationTitleSelect } from '@cbhq/cds-mobile/navigation/NavigationTitleSelect'",
     ],
     props: {
       title: figma.string('↳ title'),
@@ -157,11 +157,11 @@ figma.connect(
       type: 'with search',
     },
     imports: [
-      "import { Divider } from '@cbhq/cds-mobile/layout/Divider';",
-      "import { VStack } from '@cbhq/cds-mobile/layout/VStack';",
-      "import { HStack } from '@cbhq/cds-mobile/layout/HStack';",
-      "import { BrowserBar } from '@cbhq/cds-mobile/navigation/BrowserBar';",
-      "import { BrowserBarSearchInput } from '@cbhq/cds-mobile/navigation/BrowserBarSearchInput';",
+      "import { Divider } from '@cbhq/cds-mobile/layout/Divider'",
+      "import { VStack } from '@cbhq/cds-mobile/layout/VStack'",
+      "import { HStack } from '@cbhq/cds-mobile/layout/HStack'",
+      "import { BrowserBar } from '@cbhq/cds-mobile/navigation/BrowserBar'",
+      "import { BrowserBarSearchInput } from '@cbhq/cds-mobile/navigation/BrowserBarSearchInput'",
     ],
     props: {
       startAction: figma.boolean('show left action', {
@@ -215,10 +215,10 @@ figma.connect(
       type: 'empty',
     },
     imports: [
-      "import { TopNavBar } from '@cbhq/cds-mobile/navigation/TopNavBar';",
-      "import { Divider } from '@cbhq/cds-mobile/layout/Divider';",
-      "import { VStack } from '@cbhq/cds-mobile/layout/VStack';",
-      "import { HStack } from '@cbhq/cds-mobile/layout/HStack';",
+      "import { TopNavBar } from '@cbhq/cds-mobile/navigation/TopNavBar'",
+      "import { Divider } from '@cbhq/cds-mobile/layout/Divider'",
+      "import { VStack } from '@cbhq/cds-mobile/layout/VStack'",
+      "import { HStack } from '@cbhq/cds-mobile/layout/HStack'",
     ],
     props: {
       startAction: figma.boolean('show left action', {
@@ -270,10 +270,10 @@ figma.connect(
       type: 'stepper',
     },
     imports: [
-      "import { TopNavBar } from '@cbhq/cds-mobile/navigation/TopNavBar';",
-      "import { Divider } from '@cbhq/cds-mobile/layout/Divider';",
-      "import { VStack } from '@cbhq/cds-mobile/layout/VStack';",
-      "import { HStack } from '@cbhq/cds-mobile/layout/HStack';",
+      "import { TopNavBar } from '@cbhq/cds-mobile/navigation/TopNavBar'",
+      "import { Divider } from '@cbhq/cds-mobile/layout/Divider'",
+      "import { VStack } from '@cbhq/cds-mobile/layout/VStack'",
+      "import { HStack } from '@cbhq/cds-mobile/layout/HStack'",
     ],
     props: {
       startAction: figma.boolean('show left action', {
@@ -325,10 +325,10 @@ figma.connect(
       type: 'Market Selector',
     },
     imports: [
-      "import { TopNavBar } from '@cbhq/cds-mobile/navigation/TopNavBar';",
-      "import { Divider } from '@cbhq/cds-mobile/layout/Divider';",
-      "import { VStack } from '@cbhq/cds-mobile/layout/VStack';",
-      "import { HStack } from '@cbhq/cds-mobile/layout/HStack';",
+      "import { TopNavBar } from '@cbhq/cds-mobile/navigation/TopNavBar'",
+      "import { Divider } from '@cbhq/cds-mobile/layout/Divider'",
+      "import { VStack } from '@cbhq/cds-mobile/layout/VStack'",
+      "import { HStack } from '@cbhq/cds-mobile/layout/HStack'",
     ],
     props: {
       children: figma.children('SelectChip'),