+++ packages/mobile/src/visualizations/__figma__/ProgressCircle.figma.tsx
@@ -8,7 +8,7 @@ figma.connect(
   'https://www.figma.com/design/k5CtyJccNQUGMI5bI4lJ2g/✨-CDS-Components?node-id=64-917&m=dev',
   {
     imports: [
-      "import { ProgressCircle } from '@cbhq/cds-mobile/visualizations/ProgressCircle';",
+      "import { ProgressCircle } from '@cbhq/cds-mobile/visualizations/ProgressCircle'",
     ],
     props: {
       hideText: figma.boolean('progress label', {