+++ packages/web-visualization/src/sparkline/__figma__/Sparkline.figma.tsx
@@ -10,8 +10,8 @@ figma.connect(
   'https://www.figma.com/design/k5CtyJccNQUGMI5bI4lJ2g/✨-CDS-Components?node-id=320-15040&m=dev',
   {
     imports: [
-      "import { Sparkline } from '@cbhq/cds-web-visualization';",
-      "import { useSparklinePath } from '@cbhq/cds-common/visualizations/useSparklinePath';",
+      "import { Sparkline } from '@cbhq/cds-web-visualization'",
+      "import { useSparklinePath } from '@cbhq/cds-common/visualizations/useSparklinePath'",
     ],
     example: () => {
       const data = [20, 30, 5, 45, 0];