+++ packages/web/src/page/__figma__/PageHeader.figma.tsx
@@ -9,8 +9,8 @@ figma.connect(
   'https://www.figma.com/design/k5CtyJccNQUGMI5bI4lJ2g/✨-CDS-Components?node-id=17685%3A3171',
   {
     imports: [
-      "import { PageHeader } from '@cbhq/cds-web/page/PageHeader';",
-      "import { HStack } from '@cbhq/cds-web/layout/HStack';",
+      "import { PageHeader } from '@cbhq/cds-web/page/PageHeader'",
+      "import { HStack } from '@cbhq/cds-web/layout/HStack'",
     ],
     props: {
       start: figma.boolean('show start', {